requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
